turing/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-crispy-forms==1.14.0
django-simple-history==3.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
